+++ src/Pyro4/naming.py
@@ -297,8 +297,19 @@ class NameServerDaemon(core.Daemon):
 
 class BroadcastServer(object):
     REQUEST_NSURI = "GET_NSURI" if sys.platform == "cli" else b"GET_NSURI"
+    _pyroMultiplexOnlyReadEvents = True  # seems selectors are flipping when also using event.WRITE on a bc socket
+
+    class TransportServerAdapter(object):
+        # this adapter is used to be able to pass the BroadcastServer to Daemon.combine() to integrate the event loops.
+        def __init__(self, bcserver):
+            self.sockets = [bcserver]
+
+        def events(self, eventobjects):
+            for bc in eventobjects:
+                bc.processRequest()
 
     def __init__(self, nsUri, bchost=None, bcport=None):
+        self.transportServer = self.TransportServerAdapter(self)
         self.nsUri = nsUri
         if bcport is None:
             bcport = Pyro4.config.NS_BCPORT